Some fixes: moving fluxGen to within the loop over sources, and enabling different execution for HI & continuum cases. Also some tidying up.
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@2335 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/Analysis/simulations/trunk/simulationutilities/Continuum.cc
@@ -79,9 +79,9 @@ namespace askap {
       }
       
 
-      float Continuum::flux(float freq)
+      double Continuum::flux(double freq)
       {
-	float powerTerm = this->itsAlpha+this->itsBeta*log(freq/this->itsNuZero);
+	double powerTerm = this->itsAlpha+this->itsBeta*log(freq/this->itsNuZero);
 	return this->itsFluxZero * pow(freq/this->itsNuZero, powerTerm);
       }
